Move common code in Http1xConnectionBase and make Http1xConnectionBase implement HttpConnection interface
